fix: block.hopper: invalidate inventory change tracking when removing inventory
Fixes:
Starting to listen to recently placed double chests (when hoppers start to sleep) invalidates the previous listeners caches (same hopper). This led to hoppers sleeping with an invalidated cache, meaning later wakeup notifications from the double inventory were treated as foreign and were ignored, leading to the hopper not waking up and not reacting to changes in the double inventory.